refactor: certifica que RemoteAuthentication chama método toFormData
